refacto: Reorganize Model part
